Move doctrine extension to a better place
